database/pyproject.toml
[tool.poetry]
name = "database"
version = "0.1.0"
description = ""
authors = ["Simón <simon.garcial@udea.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
gunicorn = "^20.1.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
email-validator = "^1.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest-asyncio = "^0.20.3"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.20.0"
autopep8 = "^2.0.1"
flake8 = "^6.0.0"
black = "^23.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
